requirements.txt
azure.identity
argparse
